add: 宝用のAPIを追加
宝の内容を取得
宝の入手
ユーザが所持している宝を全て取得
